test: certificate replacement
